Fix types.

+++ primitives/src/network.rs
@@ -1,5 +1,5 @@
 use scale_info::prelude::string::String;
 
-pub type NetworkId = u8;
+pub type NetworkId = u16;
 pub type ChainName = String;
 pub type ChainNetwork = String;
